Added priority class to @incobator/azure-monitor-containers.
azure-monitor containers chart was running omsagent daemonset  without priority class defined. This opened a window for a scenario where application pods (running with the same default priority) fill up the node before omsagent is scheduled.

Such scenario is, in fact, frequently happened in the cluster
with  cluster-autoscaler running, after rapid scale-up under a condition of a lot of application unscheduled pods.

This change addresses the issue by adding a default priority 1000 to the omsagent daemonset and optionally allows to customize it.